+++ projects/com.oracle.truffle.llvm.parser/src/com/oracle/truffle/llvm/parser/metadata/debuginfo/DebugInfoFunctionProcessor.java
@@ -48,6 +48,7 @@ import com.oracle.truffle.llvm.parser.model.functions.FunctionParameter;
 import com.oracle.truffle.llvm.parser.model.symbols.constants.NullConstant;
 import com.oracle.truffle.llvm.parser.model.symbols.instructions.DbgDeclareInstruction;
 import com.oracle.truffle.llvm.parser.model.symbols.instructions.DbgValueInstruction;
+import com.oracle.truffle.llvm.parser.model.symbols.instructions.DebugTrapInstruction;
 import com.oracle.truffle.llvm.parser.model.symbols.instructions.Instruction;
 import com.oracle.truffle.llvm.parser.model.symbols.instructions.ValueInstruction;
 import com.oracle.truffle.llvm.parser.model.symbols.instructions.VoidCallInstruction;
@@ -59,6 +60,7 @@ import com.oracle.truffle.llvm.runtime.debug.scope.LLVMSourceSymbol;
 import com.oracle.truffle.llvm.runtime.debug.type.LLVMSourceType;
 import com.oracle.truffle.llvm.runtime.debug.scope.LLVMSourceLocation;
 import com.oracle.truffle.llvm.runtime.types.MetaType;
+
 import static com.oracle.truffle.llvm.parser.metadata.debuginfo.DebugInfoCache.getDebugInfo;
 
 public final class DebugInfoFunctionProcessor {
@@ -79,6 +81,8 @@ public final class DebugInfoFunctionProcessor {
     private static final int LLVM_DBG_VALUE_LOCALREF_ARGINDEX_NEW = 1;
     private static final int LLVM_DBG_VALUE_EXPR_ARGINDEX_NEW = 2;
     private static final int LLVM_DBG_VALUE_LOCALREF_ARGSIZE_NEW = 3;
+
+    private static final String LLVM_DEBUGTRAP_NAME = "@llvm.debugtrap";
 
     private final DebugInfoCache cache;
 
@@ -185,10 +189,20 @@ public final class DebugInfoFunctionProcessor {
                     case LLVM_DBG_VALUE_NAME:
                         handleDebugIntrinsic(call, false);
                         return;
+
+                    case LLVM_DEBUGTRAP_NAME:
+                        visitDebugTrap(call);
+                        return;
                 }
             }
 
             visitInstruction(call);
+        }
+
+        private void visitDebugTrap(VoidCallInstruction call) {
+            final DebugTrapInstruction trap = DebugTrapInstruction.create(call);
+            currentBlock.set(blockInstIndex, trap);
+            visitInstruction(trap);
         }
 
         private SourceVariable getVariable(VoidCallInstruction call, int index) {